+++ src/main/java/com/rcextract/minecord/ServerIdentity.java
@@ -1,38 +1,39 @@
 package com.rcextract.minecord;
 
-import java.util.Arrays;
+/*import java.util.Arrays;
 import java.util.HashSet;
-import java.util.Set;
+import java.util.Set;*/
 
 import org.apache.commons.lang.Validate;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.permissions.Permission;
 
-public class ServerIdentity implements Cloneable, ListenerHolder {
+@Deprecated
+public class ServerIdentity implements Cloneable/*, ListenerHolder*/ {
 
 	private String name;
 	private final Server server;
 	private boolean joined;
 	private Rank rank;
-	private final Set<Listener> listeners;
-	public ServerIdentity(Server server, boolean joined, Rank rank, Listener ... listeners) {
+	//private final Set<Listener> listeners;
+	public ServerIdentity(Server server, boolean joined, Rank rank/*, Listener ... listeners*/) {
 		Validate.notNull(server);
 		if (rank == null) rank = server.getRankManager().getMain();
 		this.server = server;
 		this.joined = joined;
 		this.rank = rank;
-		this.listeners = new HashSet<Listener>(Arrays.asList(listeners));
+		/*this.listeners = new HashSet<Listener>(Arrays.asList(listeners));
 		if (this.listeners.isEmpty()) this.listeners.add(new Listener(server.getMain(), true, 0));
 		validate();
-		registerAllAbsentListeners(true, 0);
+		registerAllAbsentListeners(true, 0);*/
 	}
-	private void validate() {
+	/*private void validate() {
 		boolean sameserver = true;
 		for (Listener listener : listeners) 
 			sameserver = sameserver && listener.getChannel().getServer() == server;
 		sameserver = sameserver && rank.getRankManager().getServer() == server;
 		if (!(sameserver)) throw new IllegalArgumentException();
-	}
+	}*/
 	public String getName() {
 		return name;
 	}
@@ -53,7 +54,7 @@ public class ServerIdentity implements Cloneable, ListenerHolder {
 	}
 	public void setRank(Rank rank) {
 		this.rank = rank;
-		validate();
+		//validate();
 	}
 	public void applyRank(Rank old) {
 		net.milkbowl.vault.permission.Permission pm = Minecord.getPermissionManager();
@@ -63,52 +64,6 @@ public class ServerIdentity implements Cloneable, ListenerHolder {
 				pm.playerRemove(null, player, permission.getName());
 		for (Permission permission : rank.getPermissions()) 
 			pm.playerAdd(null, player, permission.getName());
-	}
-	@Override
-	public Set<Listener> getListeners() {
-		return new HashSet<Listener>(listeners);
-	}
-	@Override
-	public Listener getListener(Channel channel) {
-		for (Listener listener : listeners) 
-			if (listener.getChannel() == channel) 
-				return listener;
-		return null;
-	}
-	@Override
-	public Set<Listener> getListeners(boolean notify) {
-		Set<Listener> listeners = new HashSet<Listener>();
-		for (Listener listener : listeners) 
-			if (listener.isNotify()) 
-				listeners.add(listener);
-		return listeners;
-	}
-	public Listener registerListener(Channel channel, boolean notify, int index) throws DuplicatedException {
-		return registerListener(new Listener(channel, notify, index));
-	}
-	public Listener registerListener(Listener listener) throws DuplicatedException {
-		if (getListener(listener.getChannel()) != null) throw new DuplicatedException();
-		if (listener.getChannel().getServer() != server) throw new IllegalArgumentException();
-		listeners.add(listener);
-		return listener;
-	}
-	public Set<Listener> registerAllAbsentListeners(boolean allNotify, int allIndex, Channel ... exceptions) {
-		Set<Listener> listeners = new HashSet<Listener>();
-		for (Channel channel : server.getChannels()) 
-			if (getListener(channel) == null && !(Arrays.asList(exceptions).contains(channel))) 
-				listeners.add(new Listener(channel, allNotify, allIndex));
-		this.listeners.addAll(listeners);
-		return listeners;
-	}
-	public void registerAllListeners(Listener ... listeners) throws DuplicatedException {
-		for (Listener listener : listeners) 
-			registerListener(listener);
-	}
-	public boolean removeListener(Listener listener, Listener main) {
-		if (!(listeners.contains(main))) throw new IllegalArgumentException();
-		if (listener == getUser().getMain()) getUser().setMain(main);
-		if (listener == main) throw new IllegalArgumentException();
-		return listeners.remove(listener);
 	}
 	public User getUser() {
 		for (User user : Minecord.getUserManager().getUsers())